Fixes on latency calculation / New Round Robin Dynamic scheduling strategy

+++ cepsim-core/src/main/scala/ca/uwo/eng/sel/cepsim/metric/LatencyMetric.scala
@@ -106,7 +106,7 @@ object LatencyMetric extends Metric {
       if (totalEvents < 0) {
         val newest = neededEntries.last
         val remainingQuantity = -totalEvents // totalEvents is already negative
-        historyTmp = historyTmp.add(Processed(newest.cloudlet, newest.time, newest.v, remainingQuantity.toInt))
+        historyTmp = historyTmp.add(Processed(newest.cloudlet, newest.time, newest.v, remainingQuantity))
       }
 
       minimumTime
